Make policy usage consistent by requiring it unspecialized
Earlier, the user could make the mistake of invoking

    safe_cast_impl<T, S, policy_log<Foobar, PolicyArg>, void>

making a double mistake in inconsistency of policy_log's template arguments
vs. safe_cast_impl arguments. In this example, the user has to make sure
that T==Foobar and PolicyArg==void.

With this change, the user only passes policy_log, without specialization,
and the library specializes it with the correct arguments.