safe_arith: basic arithmetic operations that handle overflow/underflow
Extend safe_t so that it creates a new safe object as a result of these
operations, which has the correct type to contain the result without
overflow/underflow.

Add safe_t::type typedef to as a synonym for T to be able to access the
resulting type of an arithmetic operation.

Minor macro cosmetics.