Add helpers to create a copy of safe with a new policy
While safe_t may be used directly to create a safe_t object with the
desired policy and policy_arg, these two helpers make the process much
more readable and streamlined.

Two generic policy helpers allow for setting of any policy by passing it as
a template argument and several others are shorthands for the built-in
policies.

The default policy remains policy_truncate.